feat: export `Decoder`, `Encoder` and `DecodedMessage` types from root
Fixes #1010
