+++ packages/core/src/index.ts
@@ -1,9 +1,10 @@
 export { DefaultPubSubTopic } from "./lib/constants.js";
 export { DefaultUserAgent } from "./lib/waku.js";
 
-export {
-  createEncoder,
-  createDecoder,
+export { createEncoder, createDecoder } from "./lib/message/version_0.js";
+export type {
+  Encoder,
+  Decoder,
   DecodedMessage,
 } from "./lib/message/version_0.js";
 export * as message from "./lib/message/index.js";
